Entity Relationships
Created the last entities for this LMS Spring Boot Application. Those include "Genre" and "UserRole." The first relationship is one-to-many between the genres and book tables - one genre can have many books in it. The second relationship is between the user and the user role entities - one user can have many roles and a single role can have many users in it.